refactor: move link driver to vehicle to it's own template
